Restructure and organize classes
